claude-const-maximizer/backend/templates/boilerplates/backend/code-review-assistant/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# AI and Code Analysis
openai==1.3.7
anthropic==0.7.8
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1

# Code Analysis and Security
bandit==1.7.5
safety==2.3.5
semgrep==1.45.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Code Parsing and AST
ast==0.0.2
astroid==3.0.1
tree-sitter==0.20.4
tree-sitter-python==0.20.4

# Security and Vulnerability Detection
safety==2.3.5
bandit==1.7.5
semgrep==1.45.0
snyk==1.0.0

# Performance Analysis
memory-profiler==0.61.0
psutil==5.9.6
py-spy==0.3.14

# File Processing
python-magic==0.4.27
chardet==5.2.0
gitpython==3.1.40

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
tenacity==8.2.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
